+++ MappingModule/vertex.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <stdbool.h>
 #include <string.h>
+#include <stdint.h>
 
 #include "vertex.h"
 
@@ -14,7 +15,7 @@ Vertex *Vertex_createVertex(int x, int y)
     v->x = x;
     v->y = y;
     v->visited = false;
-    for (size_t i = 0; i < 4; i++)
+    for (uint8_t i = 0; i < 4; i++)
         v->adjacencyList[i] = NULL;
 
     return v;
@@ -32,7 +33,7 @@ void Vertex_writeStrToBuff(char buff[], Vertex *v)
 
     char temp[10] = {0};
     snprintf(buff, 100, "(%d, %d): ", v->x, v->y);
-    for (size_t i = 0; i < 4; i++)
+    for (uint8_t i = 0; i < 4; i++)
     {
         if (v->adjacencyList[i] == NULL)
             continue;
